1）修改定时器方式，减少中断次数 2）发现ADS1256对时序比较敏感，不能打断读时候的时序，否则会导致错误 3）打包上位机波形显示程序 4）没有增加参数保存功能，接下来版本处理 5）8通道全开时间需要4ms时间

+++ Inc/main.h
@@ -75,7 +75,7 @@ typedef struct{
     uint8_t is_adc_drdy:1;
     uint8_t is_timer_expired:1; 
     
-    uint32_t report_interval_cout; 
+    uint16_t report_interval_cout; 
 }system_conf_t;
 
 extern system_conf_t system_conf;
